refactor(web): update after pr comments
