+++ apps/web/src/App.tsx
@@ -8,6 +8,8 @@ import { library } from '@fortawesome/fontawesome-svg-core';
 import { far } from '@fortawesome/free-regular-svg-icons';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 
+import { ProductUseCasesEnum } from '@novu/shared';
+
 import { AuthProvider } from './components/providers/AuthProvider';
 import { applyToken, getToken } from './hooks';
 import { ActivitiesPage } from './pages/activities/ActivitiesPage';
@@ -59,9 +61,8 @@ import { SSO, UserAccess, Cloud } from '@novu/design-system';
 import { BrandingForm, LayoutsListPage } from './pages/brand/tabs';
 import { TranslationRoutes } from './pages/TranslationPages';
 import { VariantsPage } from './pages/templates/components/VariantsPage';
-import { GetStartedUseCase } from './pages/get-started/GetStartedUseCase';
+import { GetStartedPage } from './pages/get-started/GetStartedPage';
 import { GetStartedTab } from './pages/get-started/layout/GetStartedTab';
-import { onboardingUseCases } from './pages/get-started/consts';
 
 library.add(far, fas);
 
@@ -171,18 +172,9 @@ function App() {
                     <Route path="create" element={<CreateTenantPage />} />
                     <Route path=":identifier" element={<UpdateTenantPage />} />
                   </Route>
-                  <Route path={ROUTES.GET_STARTED} element={<GetStartedUseCase />}>
-                    <Route path="inapp" element={<GetStartedTab onboardingUseCase={onboardingUseCases.in_app} />} />
-                    <Route
-                      path="multichannel"
-                      element={<GetStartedTab onboardingUseCase={onboardingUseCases.multi_channel} />}
-                    />
-                    <Route path="digest" element={<GetStartedTab onboardingUseCase={onboardingUseCases.digest} />} />
-                    <Route path="delay" element={<GetStartedTab onboardingUseCase={onboardingUseCases.delay} />} />
-                    <Route
-                      path="translation"
-                      element={<GetStartedTab onboardingUseCase={onboardingUseCases.translation} />}
-                    />
+                  <Route path={ROUTES.GET_STARTED} element={<GetStartedPage />}>
+                    <Route path="" element={<GetStartedTab usecase={ProductUseCasesEnum.IN_APP} />} />
+                    <Route path=":usecase" element={<GetStartedTab />} />
                   </Route>
                   <Route path={ROUTES.GET_STARTED_PREVIEW} element={<DigestPreview />} />
                   <Route path={ROUTES.QUICK_START_NOTIFICATION_CENTER} element={<NotificationCenter />} />
