+++ apps/web/src/constants/onboarding-tabs.ts
@@ -1,0 +1,7 @@
+export enum OnboardingUseCasesTabsEnum {
+  IN_APP = 'in-app',
+  DIGEST = 'digest',
+  DELAY = 'delay',
+  TRANSLATION = 'translation',
+  MULTI_CHANNEL = 'multi-channel',
+}
