+++ apps/web/src/pages/get-started/types.ts
@@ -1,0 +1,1 @@
+export type OnboardingParams = 'usecase';
